+++ exporters/stdout/stdouttrace/trace.go
@@ -6,13 +6,14 @@ package stdouttrace // import "go.opentelemetry.io/otel/exporters/stdout/stdoutt
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace/internal/counter"
 	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace/internal/x"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk"
@@ -22,8 +23,10 @@ import (
 	"go.opentelemetry.io/otel/semconv/v1.36.0/otelconv"
 )
 
-// otelComponentType is a name identifying the type of the OpenTelemetry component.
-const otelComponentType = "stdout_trace_exporter"
+// otelComponentType is a name identifying the type of the OpenTelemetry
+// component. It is not a standardized OTel component type, so it uses the
+// Go package prefixed type name to ensure uniqueness and identity.
+const otelComponentType = "go.opentelemetry.io/otel/exporters/stdout/stdouttrace.Exporter"
 
 var zeroTime time.Time
 
@@ -42,9 +45,41 @@ func New(options ...Option) (*Exporter, error) {
 		encoder:    enc,
 		timestamps: cfg.Timestamps,
 	}
-	exporter.initSelfObservability()
 
-	return exporter, nil
+	if !x.SelfObservability.Enabled() {
+		return exporter, nil
+	}
+
+	exporter.selfObservabilityEnabled = true
+	exporter.selfObservabilityAttrs = []attribute.KeyValue{
+		semconv.OTelComponentName(fmt.Sprintf("%s/%d", otelComponentType, counter.NextExporterID())),
+		semconv.OTelComponentTypeKey.String(otelComponentType),
+	}
+	s := attribute.NewSet(exporter.selfObservabilityAttrs...)
+	exporter.selfObservabilitySetOpt = metric.WithAttributeSet(s)
+
+	mp := otel.GetMeterProvider()
+	m := mp.Meter(
+		"go.opentelemetry.io/otel/exporters/stdout/stdouttrace",
+		metric.WithInstrumentationVersion(sdk.Version()),
+		metric.WithSchemaURL(semconv.SchemaURL),
+	)
+
+	var err, e error
+	if exporter.spanInflightMetric, e = otelconv.NewSDKExporterSpanInflight(m); e != nil {
+		e = fmt.Errorf("failed to create span inflight metric: %w", e)
+		err = errors.Join(err, e)
+	}
+	if exporter.spanExportedMetric, e = otelconv.NewSDKExporterSpanExported(m); e != nil {
+		e = fmt.Errorf("failed to create span exported metric: %w", e)
+		err = errors.Join(err, e)
+	}
+	if exporter.operationDurationMetric, e = otelconv.NewSDKExporterOperationDuration(m); e != nil {
+		e = fmt.Errorf("failed to create operation duration metric: %w", e)
+		err = errors.Join(err, e)
+	}
+
+	return exporter, err
 }
 
 // Exporter is an implementation of trace.SpanSyncer that writes spans to stdout.
@@ -58,59 +93,104 @@ type Exporter struct {
 
 	selfObservabilityEnabled bool
 	selfObservabilityAttrs   []attribute.KeyValue // selfObservability common attributes
+	selfObservabilitySetOpt  metric.MeasurementOption
 	spanInflightMetric       otelconv.SDKExporterSpanInflight
 	spanExportedMetric       otelconv.SDKExporterSpanExported
 	operationDurationMetric  otelconv.SDKExporterOperationDuration
 }
 
-// initSelfObservability initializes self-observability for the exporter if enabled.
-func (e *Exporter) initSelfObservability() {
-	if !x.SelfObservability.Enabled() {
-		return
+var (
+	measureAttrsPool = sync.Pool{
+		New: func() any {
+			// "component.name" + "component.type" + "error.type"
+			const n = 1 + 1 + 1
+			s := make([]attribute.KeyValue, 0, n)
+			// Return a pointer to a slice instead of a slice itself
+			// to avoid allocations on every call.
+			return &s
+		},
 	}
 
-	e.selfObservabilityEnabled = true
-	e.selfObservabilityAttrs = []attribute.KeyValue{
-		semconv.OTelComponentName(fmt.Sprintf("%s/%d", otelComponentType, nextExporterID())),
-		semconv.OTelComponentTypeKey.String(otelComponentType),
+	addOptPool = &sync.Pool{
+		New: func() any {
+			const n = 1 // WithAttributeSet
+			o := make([]metric.AddOption, 0, n)
+			return &o
+		},
 	}
 
-	mp := otel.GetMeterProvider()
-	m := mp.Meter("go.opentelemetry.io/otel/exporters/stdout/stdouttrace",
-		metric.WithInstrumentationVersion(sdk.Version()),
-		metric.WithSchemaURL(semconv.SchemaURL),
-	)
-
-	var err error
-	if e.spanInflightMetric, err = otelconv.NewSDKExporterSpanInflight(m); err != nil {
-		otel.Handle(err)
+	recordOptPool = &sync.Pool{
+		New: func() any {
+			const n = 1 // WithAttributeSet
+			o := make([]metric.RecordOption, 0, n)
+			return &o
+		},
 	}
-	if e.spanExportedMetric, err = otelconv.NewSDKExporterSpanExported(m); err != nil {
-		otel.Handle(err)
-	}
-	if e.operationDurationMetric, err = otelconv.NewSDKExporterOperationDuration(m); err != nil {
-		otel.Handle(err)
-	}
-}
+)
 
 // ExportSpans writes spans in json format to stdout.
 func (e *Exporter) ExportSpans(ctx context.Context, spans []trace.ReadOnlySpan) (err error) {
+	var success int64
 	if e.selfObservabilityEnabled {
 		count := int64(len(spans))
 
-		e.spanInflightMetric.Add(context.Background(), count, e.selfObservabilityAttrs...)
+		addOpt := addOptPool.Get().(*[]metric.AddOption)
+		defer func() {
+			*addOpt = (*addOpt)[:0]
+			addOptPool.Put(addOpt)
+		}()
+
+		*addOpt = append(*addOpt, e.selfObservabilitySetOpt)
+
+		e.spanInflightMetric.Inst().Add(ctx, count, *addOpt...)
 		defer func(starting time.Time) {
-			// additional attributes for self-observability,
-			// only spanExportedMetric and operationDurationMetric are supported
-			addAttrs := make([]attribute.KeyValue, len(e.selfObservabilityAttrs), len(e.selfObservabilityAttrs)+1)
-			copy(addAttrs, e.selfObservabilityAttrs)
+			e.spanInflightMetric.Inst().Add(ctx, -count, *addOpt...)
+
+			// Record the success and duration of the operation.
+			//
+			// Do not exclude 0 values, as they are valid and indicate no spans
+			// were exported which is meaningful for certain aggregations.
+			e.spanExportedMetric.Inst().Add(ctx, success, *addOpt...)
+
+			mOpt := e.selfObservabilitySetOpt
 			if err != nil {
-				addAttrs = append(addAttrs, semconv.ErrorType(err))
+				// additional attributes for self-observability,
+				// only spanExportedMetric and operationDurationMetric are supported.
+				attrs := measureAttrsPool.Get().(*[]attribute.KeyValue)
+				defer func() {
+					*attrs = (*attrs)[:0] // reset the slice for reuse
+					measureAttrsPool.Put(attrs)
+				}()
+				*attrs = append(*attrs, e.selfObservabilityAttrs...)
+				*attrs = append(*attrs, semconv.ErrorType(err))
+
+				// Do not inefficiently make a copy of attrs by using
+				// WithAttributes instead of WithAttributeSet.
+				set := attribute.NewSet(*attrs...)
+				mOpt = metric.WithAttributeSet(set)
+
+				// Reset addOpt with new attribute set.
+				*addOpt = append((*addOpt)[:0], mOpt)
+
+				e.spanExportedMetric.Inst().Add(
+					ctx,
+					count-success,
+					*addOpt...,
+				)
 			}
 
-			e.spanInflightMetric.Add(context.Background(), -count, e.selfObservabilityAttrs...)
-			e.spanExportedMetric.Add(context.Background(), count, addAttrs...)
-			e.operationDurationMetric.Record(context.Background(), time.Since(starting).Seconds(), addAttrs...)
+			recordOpt := recordOptPool.Get().(*[]metric.RecordOption)
+			defer func() {
+				*recordOpt = (*recordOpt)[:0]
+				recordOptPool.Put(recordOpt)
+			}()
+
+			*recordOpt = append(*recordOpt, mOpt)
+			e.operationDurationMetric.Inst().Record(
+				ctx,
+				time.Since(starting).Seconds(),
+				*recordOpt...,
+			)
 		}(time.Now())
 	}
 
@@ -145,11 +225,13 @@ func (e *Exporter) ExportSpans(ctx context.Context, spans []trace.ReadOnlySpan) 
 		}
 
 		// Encode span stubs, one by one
-		if err := e.encoder.Encode(stub); err != nil {
-			return err
+		if e := e.encoder.Encode(stub); e != nil {
+			err = errors.Join(err, fmt.Errorf("failed to encode span %d: %w", i, e))
+			continue
 		}
+		success++
 	}
-	return nil
+	return err
 }
 
 // Shutdown is called to stop the exporter, it performs no action.
@@ -170,12 +252,4 @@ func (e *Exporter) MarshalLog() any {
 		Type:           "stdout",
 		WithTimestamps: e.timestamps,
 	}
-}
-
-var exporterIDCounter atomic.Int64
-
-// nextExporterID returns a new unique ID for an exporter.
-// the starting value is 0, and it increments by 1 for each call.
-func nextExporterID() int64 {
-	return exporterIDCounter.Add(1) - 1
 }