+++ sdk/trace/span.go
@@ -497,8 +497,8 @@ func (s *recordingSpan) End(options ...trace.SpanEndOption) {
 	}
 	s.mu.Unlock()
 
-	defer func() {
-		if s.tracer.selfObservabilityEnabled {
+	if s.tracer.selfObservabilityEnabled {
+		defer func() {
 			// Determine the sampling result and create the corresponding attribute.
 			var attrSamplingResult attribute.KeyValue
 			if s.spanContext.IsSampled() {
@@ -509,9 +509,12 @@ func (s *recordingSpan) End(options ...trace.SpanEndOption) {
 				attrSamplingResult = s.tracer.spanLiveMetric.AttrSpanSamplingResult(otelconv.SpanSamplingResultRecordOnly)
 			}
 
-			s.tracer.spanLiveMetric.Add(context.Background(), -1, attrSamplingResult)
-		}
-	}()
+			// Add the span to the context to ensure the metric is recorded
+			// with the correct span context.
+			ctx := trace.ContextWithSpan(context.Background(), s)
+			s.tracer.spanLiveMetric.Add(ctx, -1, attrSamplingResult)
+		}()
+	}
 
 	sps := s.tracer.provider.getSpanProcessors()
 	if len(sps) == 0 {